+++ pyxelrest/_exceptions.py
@@ -1,8 +1,12 @@
-class InvalidOpenAPIDefinition(Exception):
+class PyxelRestException(Exception):
+    pass
+
+
+class InvalidOpenAPIDefinition(PyxelRestException):
     """ Invalid OpenAPI Definition. """
 
     def __init__(self, message: str, *args, **kwargs):  # real signature unknown
-        Exception.__init__(self, "Invalid Definition: " + message)
+        PyxelRestException.__init__(self, "Invalid Definition: " + message)
 
 
 class OpenAPIVersionNotProvided(InvalidOpenAPIDefinition):
@@ -19,20 +23,20 @@ class UnsupportedOpenAPIVersion(InvalidOpenAPIDefinition):
         InvalidOpenAPIDefinition.__init__(self, f"Version {version} not supported.")
 
 
-class MandatoryPropertyNotProvided(Exception):
+class MandatoryPropertyNotProvided(PyxelRestException):
     """ Mandatory property not provided. """
 
     def __init__(self, section: str, property_name: str, *args, **kwargs):
-        Exception.__init__(
+        PyxelRestException.__init__(
             self, f'"{section}" configuration section must provide "{property_name}".'
         )
 
 
-class DuplicatedParameters(Exception):
+class DuplicatedParameters(PyxelRestException):
     """ Method contains duplicated parameters. """
 
     def __init__(self, method: str, *args, **kwargs):
-        Exception.__init__(
+        PyxelRestException.__init__(
             self,
             f'"{method["operationId"]}" parameters are not unique per location: {method["parameters"]}.',
         )